pyproject.toml
[project]
name = "app"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "chex>=0.1.91",
    "flax>=0.12.4",
    "jax>=0.9.0.1",
    "jaxlib>=0.9.0.1",
    "matplotlib>=3.10.8",
    "numba>=0.64.0",
    "optax>=0.2.6",
    "pandapower>=3.4.0",
    "pyyaml>=6.0.3",
    "tabulate>=0.9.0",
    "tqdm>=4.67.3",
    "uvloop>=0.22.1; sys_platform != 'win32'",
]
[tool.uv]
override-dependencies = [
    "uvloop ; sys_platform != 'win32'"
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
